pyproject.toml
[project]
name = "esperanto-lm"
version = "0.1.0"
description = "Small LLaMA-style causal language model trained on Esperanto Wikipedia"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.3.0",
    "transformers>=4.45.0",
    "datasets>=2.20.0",
    "tokenizers>=0.19.0",
    "accelerate>=0.30.0",
    "trl>=0.9.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "wandb>=0.17.0",
]

[project.scripts]
train = "esperanto_lm.train:main"
train-tokenizer = "scripts.train_tokenizer:main"
download-data = "scripts.download_data:main"
generate = "scripts.generate:main"
download-hplt = "scripts.download_hplt:main"
download-gutenberg = "scripts.download_gutenberg:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/esperanto_lm", "scripts"]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }

[tool.pytest.ini_options]
testpaths = ["tests"]
